Implement syscalls to open and close files
The syscalls are passed through to the host, but virtual file
descriptors are returned to the simulated program. A table is used to
map these virtual file descriptors onto the real file descriptors when
they're passed to syscalls that use them (such as ioctl and writev).

The openat() syscall currently blacklists filepaths that start with
/dev/, /proc/ and /sys/.